Update readme for passport-pgt-client
Extended the installation section in the readme file of the passport-pgt-client package. Added comments to guide the user through the installation and configuration steps. Also, simplified the override process by advising users to just update the published config file. The changes aim to provide clearer instructions for package setup and use.