Update passport URL handling in Passport PGT Client package
Remove the fallback to 'app.url' from the 'passport_server' configuration variable and change return type of getPassportUrl and getBaseUrl methods to string|null in PassportPgtClient service.

This change updates the way we handle the passport URL in the pgt-client package. The fallback to the application URL in the configuration was removed, making the environment variable the exclusive source for this data. We also had to adjust the return types of the getPassportUrl and getBaseUrl methods to reflect this change, as it can now return null if the environment variable is not set.